REFACTOR - 共通処理部分を切り出し、mixinで読み込み共通化 またフォルダも統一し、管理しやすく
